Revert "Revert "Refactor States To Allow for Single Cached Hasher (#9922)""
This reverts commit 928e05ead084c29995ef150b4a8b204d57a51bdd.

# Conflicts:
#	beacon-chain/state/v3/BUILD.bazel
#	beacon-chain/state/v3/field_roots.go
#	beacon-chain/state/v3/state_trie.go
