+++ beacon-chain/state/fieldtrie/BUILD.bazel
@@ -32,7 +32,6 @@ go_test(
         "//beacon-chain/state/custom-types:go_default_library",
         "//beacon-chain/state/stateutil:go_default_library",
         "//beacon-chain/state/types:go_default_library",
-        "//beacon-chain/state/v1:go_default_library",
         "//config/params:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//testing/assert:go_default_library",
