+++ beacon-chain/state/stateutil/validator_root.go
@@ -88,9 +88,9 @@ func Uint64ListRootWithRegistryLimit(balances []uint64) ([32]byte, error) {
 	return ssz.MixInLength(balancesRootsRoot, balancesLengthRoot), nil
 }
 
-// ValidatorEncKey returns the encoded key in bytes of input `validator`,
+// validatorEncKey returns the encoded key in bytes of input `validator`,
 // the returned key bytes can be used for caching purposes.
-func ValidatorEncKey(validator *ethpb.Validator) []byte {
+func validatorEncKey(validator *ethpb.Validator) []byte {
 	if validator == nil {
 		return nil
 	}
